feat: braking out thei am infra, dropping the service for the gitcoin domain
